Proof of Concept: simplifying UI of nodes
Proof of concept simplifiying pages & paramaters of nodes.

Blender shows ports differently on nodes (input on the bottom left,
output on top right) as Maya. Reading the ARGS files (a XML derivat), it
looks like that Maya organises these ports in pages. Because of the
behaviour of Blender (consuming ports differently as described before)
it can be that pages are empty or a page contains only a single
parameter. This is a first aproach to disable the (useless) drawing of
empty pages on a node, or the drawing of single parameters in pages on a
node which have to be first unfolded (to much boring clicks).

This works fine now on PxrBumP, PxrTexture

But should to be also implemented on :
aaOceanPrManShader, PxrMarschnerHair, PxrBump, PxrBakePointCloud,
PxrNormalMap, PxrRemap, PxrVolume, PxrDirt, PxrDispScalarLayer,
PxrLayerSurface, PxrDispVectorLayer, PxrFractalize, PxrLayeredTexture,
PxrMultiTexture, PxrGeometricAOVs, PxrMatteID, PxrProjector,
PxrRoundCube and PxrLayerMixer